domeeni testid, ServiceForPerson servicetest kukub läbi

+++ Domain/Others/EquipmentType.cs
@@ -4,6 +4,6 @@ using Abc.Domain.Common;
 namespace Abc.Domain.Others {
     public sealed class EquipmentType : DefinedEntity<EquipmentTypeData> {
         public EquipmentType(EquipmentTypeData d) : base(d) { }
-        public double AmountAvailable => Data?.AmountAvailable ?? UnspecifiedInteger;
+        public int AmountAvailable => Data?.AmountAvailable ?? UnspecifiedInteger;
     }
 }
